Create separate module for factories
