+++ tests/factory_fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .lease import *
